+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # middy-store
 
+## 0.2.1
+
+### Patch Changes
+
+- fix: calculate lambda sync and async sizes correctly
+
 ## 0.2.0
 
 ### Minor Changes