+++ packages/store-dynamodb/CHANGELOG.md
@@ -1,5 +1,14 @@
 # middy-store-dynamodb
 
+## 0.1.1
+
+### Patch Changes
+
+- fix: calculate lambda sync and async sizes correctly
+
+- Updated dependencies []:
+  - middy-store@0.2.1
+
 ## 0.1.0
 
 ### Minor Changes